+++ modules/armies/hooks/index.ts
@@ -1,0 +1,1 @@
+export { default as useUnitTypes } from './useUnitTypes'